// frontendOld/app/types/models.ts
export interface Uyezd {
  id?: number
  uyezd?: string
}

export interface Volost {
  id?: number
  volost?: string
  uyezd?: Uyezd | null
}

export interface Place {
  id?: number
  place?: string
  volost?: Volost | null
}

export interface Landowner {
  id?: number
  landowner?: string
  place?: Place | null
}

export interface FamilyStatus {
  id?: number
  familyStatus?: string
}

export interface SocialStatus {
  id?: number
  socialStatus?: string
}

export interface FuzzyDate {
  id?: number
  exactDate?: string
  startDate?: string
  endDate?: string
  description?: string
}

export interface Person {
  id?: number
  firstName?: string
  lastName?: string
  middleName?: string
  gender?: "MALE" | "FEMALE"
  birthDate?: FuzzyDate
  deathDate?: FuzzyDate
  uyezd?: Uyezd | null
  volost?: Volost | null
  place?: Place | null
  socialStatus?: SocialStatus | null
  spouse?: Person
  father?: Person
  mother?: Person
  children?: Person[]
}